postprocess: allow using dcraw_emu from Debian
As `dcraw_emu` is an example program, Debian doesn't install it to
`/usr/bin`, but to `/usr/lib/libraw` instead, meaning it can't be found
the usual way.

This patch checks for `/usr/lib/libraw/dcraw_emu` if `command` can't
find it. Additionnally, cascaded if's are replaced with if..elif
statements, so we stop searching as soon as a matching program
(preferably `dcraw_emu`) is found.