Drop support for HDR+ stacker
The current HDR stacker is incredibly buggy and was only
added so it can be quickly be tested by building/installing it
locally. Some distributions have actually packaged this buggy mess
and made it a dependency of megapixes meaning the whole experience
is broken for users.

This drops support for trying to use stack_frames by default so
hdr-plus being installed for users doesn't break/corrupt photos.