Only disable flash when flashing is enabled
If the flash was disabled and screne flash is used it will try to set
the screen brightness back to the original value. Because the check for
mp_flash_enable and mp_flash_disable is different there's cases where
enable wasn't run but disable is, causing it to use the initial zero
brightness as the restore value.