Capture at least 2 frames
Current post processing scripts assume at least 2 frames are captured.
With the rear pinephone camera the automatic gain control will always
raise the gain high enough that the frame count will be 2 or higher. For
the front camera the gain might be zero causing the burst to be 1 frame
long.